app/requirements.txt
numpy
scikit-learn
pandas
tensorflow
joblib
psycopg2
streamlit
plotly
websocket-client
kafka-python
